requirements.txt
numpy==1.19.5
tensorflow-gpu==2.6.0
flask==2.2.3
matplotlib==3.5.3
py-opencv==4.6.0
opencv-python==4.6.0
